requirements.txt
picamera2
pyserial
lmxl
requests
mariadb-server
mysql-connector-python